feat: add error when id entered via params is not a number

+++ src/controllers/cardsController.ts
@@ -23,6 +23,13 @@ export async function getCardById(req: Request, res: Response) {
     const { user } = res.locals;
     const { id } = req.params;
 
+    if(isNaN(Number(id))) {
+        throw {
+            type: "unprocessable_entity",
+            message: "The ID must be a number."
+        };
+    }
+
     const card = await cardsService.getCardById(user.id, Number(id));
 
     res.status(200).send(card);
@@ -31,6 +38,13 @@ export async function getCardById(req: Request, res: Response) {
 export async function deleteCard(req: Request, res: Response) {
     const { user } = res.locals;
     const { id } = req.params;
+
+    if(isNaN(Number(id))) {
+        throw {
+            type: "unprocessable_entity",
+            message: "The ID must be a number."
+        };
+    }
 
     await cardsService.deleteCard(user.id, Number(id));
